testing/requirements.txt
async-timeout==4.0.3
python-dotenv==1.0.0
redis==5.0.1
